+++ lib/plug/accesslog/default_formatter/remote_ip_address.ex
@@ -4,12 +4,10 @@ defmodule Plug.AccessLog.DefaultFormatter.RemoteIPAddress do
   """
 
   @doc """
-  Appends to log output.
+  Formats the log output.
   """
-  @spec append(String.t, Plug.Conn.t) :: String.t
-  def append(message, conn), do: message <> remote_ip_address(conn)
-
-  defp remote_ip_address(conn) do
+  @spec format(Plug.Conn.t) :: iodata
+  def format(conn) do
     conn.remote_ip
     |> :inet_parse.ntoa()
     |> to_string()